Split out get-clips into a different file from db.ts
